Refactor SetName to SetNameModule
