+++ EvoSC.sln
@@ -44,7 +44,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EvoSC.Manialinks", "src\Evo
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EvoSC.Manialinks.Tests", "tests\EvoSC.Manialinks.Tests\EvoSC.Manialinks.Tests.csproj", "{BFE27507-5B63-44B9-87EE-D2E46989F3CE}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SetName", "src\Modules\SetName\SetName.csproj", "{568D81FE-858A-4052-B59B-9381E0FE604C}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SetNameModule", "src\Modules\SetNameModule\SetNameModule.csproj", "{568D81FE-858A-4052-B59B-9381E0FE604C}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FastestCpModule", "src\Modules\FastestCpModule\FastestCpModule.csproj", "{9E1335F9-6C39-4B3F-9CEB-A65EEDDF798D}"
 EndProject
