+++ src/Modules/SetNameModule/Controllers/SetNameController.cs
@@ -1,10 +1,10 @@
 ﻿using EvoSC.Common.Controllers.Attributes;
 using EvoSC.Common.Interfaces.Localization;
 using EvoSC.Manialinks;
-using EvoSC.Modules.Official.SetName.Interfaces;
-using EvoSC.Modules.Official.SetName.Models;
+using EvoSC.Modules.Official.SetNameModule.Interfaces;
+using EvoSC.Modules.Official.SetNameModule.Models;
 
-namespace EvoSC.Modules.Official.SetName.Controllers;
+namespace EvoSC.Modules.Official.SetNameModule.Controllers;
 
 [Controller]
 public class SetNameController(ISetNameService setNameService, Locale locale) : ManialinkController
@@ -13,12 +13,12 @@ public class SetNameController(ISetNameService setNameService, Locale locale) : 
     {
         if (!IsModelValid)
         {
-            await ShowAsync(Context.Player, "SetName.EditName", new {input.Nickname, Locale = locale});
+            await ShowAsync(Context.Player, "SetNameModule.EditName", new {input.Nickname, Locale = locale});
             return;
         }
 
         await setNameService.SetNicknameAsync(Context.Player, input.Nickname);
-        await HideAsync(Context.Player, "SetName.EditName");
+        await HideAsync(Context.Player, "SetNameModule.EditName");
 
         Context.AuditEvent
             .Success()
@@ -26,5 +26,5 @@ public class SetNameController(ISetNameService setNameService, Locale locale) : 
             .HavingProperties(new {OldNickname = Context.Player.NickName, NewNickname = input.Nickname});
     }
 
-    public Task CancelAsync() => HideAsync(Context.Player, "SetName.EditName");
+    public Task CancelAsync() => HideAsync(Context.Player, "SetNameModule.EditName");
 }
