+++ src/Modules/SetNameModule/SetNameModule.csproj
@@ -4,7 +4,7 @@
         <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <RootNamespace>EvoSC.Modules.Official.SetName</RootNamespace>
+        <RootNamespace>EvoSC.Modules.Official.SetNameModule</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
